Modify float to double for percentage attribut
git-svn-id: https://gsphere.genoscope.cns.fr/svn/NGL/trunk/ngl@3984 d0e871a8-0b98-4e59-9d77-5ea3c331444f

+++ lib-ngl-common/app/models/laboratory/experiment/instance/ContainerUsed.java
@@ -30,7 +30,7 @@ public class ContainerUsed implements IValidation{
 	
 	public List<Content> contents;
 	
-	public Float percentage;
+	public Double percentage;
 	// Proprietes a renseigner en fonction du type d'experiment ou d'instrument
 	public Map<String,PropertyValue> experimentProperties;
 	public Map<String,PropertyValue> instrumentProperties;
